requirements.txt
# Backend requirements
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.5.2
pillow==10.1.0
python-multipart==0.0.6
openai==1.3.0
requests==2.31.0

# Frontend requirements
streamlit==1.29.0



# langchain==0.3.13
# langchain-core==0.3.28
# langchain-community==0.3.13
# langchain-openai==0.2.14
# crewai==0.30.11
# streamlit==1.32.2
# openai==1.14.3
# pydantic==2.7.0
# faiss-cpu==1.8.0
